Remove some hacks in build file generator
Overall this change is a rename of few files.

linalg module has been split in 2 parts (common and enc) and greatly simplified.
