+++ lib/jxl/decode_test.cc
@@ -32,6 +32,7 @@
 #include "lib/jxl/enc_external_image.h"
 #include "lib/jxl/enc_file.h"
 #include "lib/jxl/enc_icc_codec.h"
+#include "lib/jxl/enc_progressive_split.h"
 #include "lib/jxl/encode_internal.h"
 #include "lib/jxl/fields.h"
 #include "lib/jxl/frame_header.h"
@@ -39,7 +40,6 @@
 #include "lib/jxl/icc_codec.h"
 #include "lib/jxl/image_metadata.h"
 #include "lib/jxl/jpeg/enc_jpeg_data.h"
-#include "lib/jxl/progressive_split.h"
 #include "lib/jxl/test_utils.h"
 #include "lib/jxl/testdata.h"
 #include "lib/jxl/toc.h"
