+++ lib/jxl/enc_cache.h
@@ -22,11 +22,11 @@
 #include "lib/jxl/enc_ans.h"
 #include "lib/jxl/enc_heuristics.h"
 #include "lib/jxl/enc_params.h"
+#include "lib/jxl/enc_progressive_split.h"
 #include "lib/jxl/frame_header.h"
 #include "lib/jxl/image.h"
 #include "lib/jxl/image_bundle.h"
 #include "lib/jxl/passes_state.h"
-#include "lib/jxl/progressive_split.h"
 #include "lib/jxl/quant_weights.h"
 #include "lib/jxl/quantizer.h"
