+++ lib/jxl/enc_gaborish.cc
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-#include "lib/jxl/gaborish.h"
+#include "lib/jxl/enc_gaborish.h"
 
 #include <stddef.h>
