+++ lib/jxl/enc_noise.cc
@@ -16,9 +16,9 @@
 #include "lib/jxl/base/compiler_specific.h"
 #include "lib/jxl/chroma_from_luma.h"
 #include "lib/jxl/convolve.h"
+#include "lib/jxl/enc_optimize.h"
 #include "lib/jxl/image_ops.h"
 #include "lib/jxl/opsin_params.h"
-#include "lib/jxl/optimize.h"
 
 namespace jxl {
 namespace {
