+++ lib/jxl/enc_optimize.cc
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-#include "lib/jxl/optimize.h"
+#include "lib/jxl/enc_optimize.h"
 
 #include <algorithm>
