+++ lib/jxl/enc_optimize.h
