+++ lib/jxl/enc_progressive_split.cc
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-#include "lib/jxl/progressive_split.h"
+#include "lib/jxl/enc_progressive_split.h"
 
 #include <string.h>
