+++ lib/lib.gni
@@ -171,11 +171,11 @@ libjxl_dec_sources = [
     "jxl/jpeg/jpeg_data.h",
     "jxl/jxl_inspection.h",
     "jxl/lehmer_code.h",
-    "jxl/linalg.h",
     "jxl/loop_filter.cc",
     "jxl/loop_filter.h",
     "jxl/luminance.cc",
     "jxl/luminance.h",
+    "jxl/matrix_ops.h",
     "jxl/memory_manager_internal.cc",
     "jxl/memory_manager_internal.h",
     "jxl/modular/encoding/context_predict.h",
@@ -302,6 +302,8 @@ libjxl_enc_sources = [
     "jxl/enc_file.h",
     "jxl/enc_frame.cc",
     "jxl/enc_frame.h",
+    "jxl/enc_gaborish.cc",
+    "jxl/enc_gaborish.h",
     "jxl/enc_gamma_correct.h",
     "jxl/enc_group.cc",
     "jxl/enc_group.h",
@@ -309,20 +311,28 @@ libjxl_enc_sources = [
     "jxl/enc_heuristics.h",
     "jxl/enc_huffman.cc",
     "jxl/enc_huffman.h",
+    "jxl/enc_huffman_tree.cc",
+    "jxl/enc_huffman_tree.h",
     "jxl/enc_icc_codec.cc",
     "jxl/enc_icc_codec.h",
     "jxl/enc_image_bundle.cc",
     "jxl/enc_image_bundle.h",
     "jxl/enc_jxl_skcms.h",
+    "jxl/enc_linalg.cc",
+    "jxl/enc_linalg.h",
     "jxl/enc_modular.cc",
     "jxl/enc_modular.h",
     "jxl/enc_noise.cc",
     "jxl/enc_noise.h",
+    "jxl/enc_optimize.cc",
+    "jxl/enc_optimize.h",
     "jxl/enc_params.h",
     "jxl/enc_patch_dictionary.cc",
     "jxl/enc_patch_dictionary.h",
     "jxl/enc_photon_noise.cc",
     "jxl/enc_photon_noise.h",
+    "jxl/enc_progressive_split.cc",
+    "jxl/enc_progressive_split.h",
     "jxl/enc_quant_weights.cc",
     "jxl/enc_quant_weights.h",
     "jxl/enc_splines.cc",
@@ -336,17 +346,12 @@ libjxl_enc_sources = [
     "jxl/enc_xyb.h",
     "jxl/encode.cc",
     "jxl/encode_internal.h",
-    "jxl/gaborish.cc",
-    "jxl/gaborish.h",
-    "jxl/huffman_tree.cc",
-    "jxl/huffman_tree.h",
     "jxl/jpeg/enc_jpeg_data.cc",
     "jxl/jpeg/enc_jpeg_data.h",
     "jxl/jpeg/enc_jpeg_data_reader.cc",
     "jxl/jpeg/enc_jpeg_data_reader.h",
     "jxl/jpeg/enc_jpeg_huffman_decode.cc",
     "jxl/jpeg/enc_jpeg_huffman_decode.h",
-    "jxl/linalg.cc",
     "jxl/modular/encoding/enc_debug_tree.cc",
     "jxl/modular/encoding/enc_debug_tree.h",
     "jxl/modular/encoding/enc_encoding.cc",
@@ -361,10 +366,6 @@ libjxl_enc_sources = [
     "jxl/modular/transform/enc_squeeze.h",
     "jxl/modular/transform/enc_transform.cc",
     "jxl/modular/transform/enc_transform.h",
-    "jxl/optimize.cc",
-    "jxl/optimize.h",
-    "jxl/progressive_split.cc",
-    "jxl/progressive_split.h",
 ]
 
 libjxl_gbench_sources = [
@@ -394,13 +395,15 @@ libjxl_tests_sources = [
     "jxl/dct_test.cc",
     "jxl/decode_test.cc",
     "jxl/enc_external_image_test.cc",
+    "jxl/enc_gaborish_test.cc",
+    "jxl/enc_linalg_test.cc",
+    "jxl/enc_optimize_test.cc",
     "jxl/enc_photon_noise_test.cc",
     "jxl/encode_test.cc",
     "jxl/entropy_coder_test.cc",
     "jxl/fast_dct_test.cc",
     "jxl/fast_math_test.cc",
     "jxl/fields_test.cc",
-    "jxl/gaborish_test.cc",
     "jxl/gamma_correct_test.cc",
     "jxl/gauss_blur_test.cc",
     "jxl/gradient_test.cc",
@@ -410,11 +413,9 @@ libjxl_tests_sources = [
     "jxl/image_ops_test.cc",
     "jxl/jxl_test.cc",
     "jxl/lehmer_code_test.cc",
-    "jxl/linalg_test.cc",
     "jxl/modular_test.cc",
     "jxl/opsin_image_test.cc",
     "jxl/opsin_inverse_test.cc",
-    "jxl/optimize_test.cc",
     "jxl/padded_bytes_test.cc",
     "jxl/passes_test.cc",
     "jxl/patch_dictionary_test.cc",
